Improved site structure and compartmentalization
