+++ src/lib/components/ListObservationUnits.svelte
@@ -1,0 +1,10 @@
+<script lang="ts">
+    export let obsUnits: ObservationUnit[];
+</script>
+
+<ul>
+    {#each obsUnits as unit}
+        <li>{unit.name}</li>
+    {/each}
+</ul>
+
